Move AddButton to common components folder
