+++ src/pages/feedbacks/components/AddButton/index.ts
@@ -1,0 +1,1 @@
+export { AddButton as default } from './AddButton';
